added logic to calculate response rate
